# server/openapi.yaml
openapi: 3.0.3
info:
  title: UrbanFlow Mobility API
  version: 1.0.0
  description: API backend UrbanFlow Mobility pour l'authentification, les lieux IDF Mobilités et les itinéraires.
servers:
  - url: http://localhost:3000
    description: Développement local
tags:
  - name: Health
    description: État de l'API
  - name: Auth
    description: Authentification par cookie httpOnly
  - name: Carbon
    description: Historique carbone des trajets terminés
  - name: IDF Mobilités
    description: Recherche de lieux et calcul d'itinéraires
paths:
  /api/health:
    get:
      tags:
        - Health
      summary: Vérifier que l'API répond.
      responses:
        '200':
          description: API disponible.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /api/auth/register:
    post:
      tags:
        - Auth
      summary: Créer un compte et ouvrir une session.
      description: Requiert un jeton CSRF transmis dans l'en-tête `x-csrf-token`.
      security:
        - csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthCredentials'
      responses:
        '201':
          description: Compte créé.
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Cookie httpOnly `urbanflow_auth`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Un compte existe déjà avec cet email.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/CsrfError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/auth/login:
    post:
      tags:
        - Auth
      summary: Ouvrir une session.
      description: Requiert un jeton CSRF transmis dans l'en-tête `x-csrf-token`.
      security:
        - csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthCredentials'
      responses:
        '200':
          description: Session ouverte.
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Cookie httpOnly `urbanflow_auth`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/CsrfError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/auth/logout:
    post:
      tags:
        - Auth
      summary: Fermer la session courante.
      description: Requiert un jeton CSRF transmis dans l'en-tête `x-csrf-token`.
      security:
        - csrfHeader: []
      responses:
        '204':
          description: Session fermée.
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Cookie `urbanflow_auth` vidé.
        '403':
          $ref: '#/components/responses/CsrfError'
  /api/auth/csrf:
    get:
      tags:
        - Auth
      summary: Obtenir un jeton CSRF.
      responses:
        '200':
          description: Jeton CSRF créé et posé dans le cookie `urbanflow_csrf`.
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Cookie lisible `urbanflow_csrf`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CsrfTokenResponse'
  /api/auth/me:
    get:
      tags:
        - Auth
      summary: Lire l'utilisateur connecté.
      security:
        - cookieAuth: []
        - {}
      responses:
        '200':
          description: Utilisateur connecté, ou aucun utilisateur si la session est absente ou invalide.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponse'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      tags:
        - Auth
      summary: Supprimer le compte connecté.
      description: Requiert un jeton CSRF transmis dans l'en-tête `x-csrf-token`.
      security:
        - cookieAuth: []
          csrfHeader: []
      responses:
        '204':
          description: Compte supprimé et session fermée.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/CsrfError'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/carbon/completed-journeys:
    get:
      tags:
        - Carbon
      summary: Lire les trajets terminés du compte connecté.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Trajets terminés synchronisés.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompletedJourneysResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags:
        - Carbon
      summary: Enregistrer ou mettre à jour un trajet terminé.
      description: Requiert un jeton CSRF transmis dans l'en-tête `x-csrf-token`.
      security:
        - cookieAuth: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompletedJourney'
      responses:
        '201':
          description: Trajet terminé enregistré.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompletedJourneyResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/CsrfError'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/idfm/nearby-stations:
    get:
      tags:
        - IDF Mobilités
      summary: Rechercher les stations proches d'une coordonnée.
      parameters:
        - $ref: '#/components/parameters/Lon'
        - $ref: '#/components/parameters/Lat'
        - name: distance
          in: query
          required: false
          schema:
            type: integer
            default: 900
            minimum: 100
            maximum: 3000
          description: Rayon en mètres.
        - name: count
          in: query
          required: false
          schema:
            type: integer
            default: 30
            minimum: 1
            maximum: 80
          description: Nombre maximal de stations.
      responses:
        '200':
          description: Stations proches.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NearbyStationsResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '502':
          $ref: '#/components/responses/IdfmUnavailable'
        '503':
          $ref: '#/components/responses/IdfmMissingToken'
        '504':
          $ref: '#/components/responses/IdfmTimeout'
  /api/idfm/disruptions:
    get:
      tags:
        - IDF Mobilités
      summary: Rechercher les perturbations en cours.
      description: Renvoie les interruptions puis les perturbations sur metro, RER, lignes rapides et tram. Les bus sont exclus.
      parameters:
        - name: count
          in: query
          required: false
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 200
          description: Nombre maximal de rapports trafic demandés à IDF Mobilités.
      responses:
        '200':
          description: Perturbations trouvées.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DisruptionsResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '502':
          $ref: '#/components/responses/IdfmUnavailable'
        '503':
          $ref: '#/components/responses/IdfmMissingToken'
        '504':
          $ref: '#/components/responses/IdfmTimeout'
  /api/idfm/bike-stations:
    get:
      tags:
        - IDF Mobilités
      summary: Rechercher les bornes Vélib proches d'une coordonnée.
      parameters:
        - $ref: '#/components/parameters/Lon'
        - $ref: '#/components/parameters/Lat'
        - name: distance
          in: query
          required: false
          schema:
            type: integer
            default: 1500
            minimum: 100
            maximum: 5000
          description: Rayon en mètres.
        - name: count
          in: query
          required: false
          schema:
            type: integer
            default: 5
            minimum: 1
            maximum: 20
          description: Nombre maximal de bornes.
        - name: availability
          in: query
          required: false
          schema:
            type: string
            enum:
              - bikes
              - docks
            default: bikes
          description: Filtre sur les vélos disponibles ou les bornettes libres.
      responses:
        '200':
          description: Bornes Vélib proches.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BikeStationsResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '502':
          $ref: '#/components/responses/IdfmUnavailable'
        '504':
          $ref: '#/components/responses/IdfmTimeout'
  /api/idfm/places:
    get:
      tags:
        - IDF Mobilités
      summary: Rechercher des lieux pour l'autocomplétion.
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Texte recherché. Moins de deux caractères renvoie une liste vide.
        - name: count
          in: query
          required: false
          schema:
            type: integer
            default: 8
            minimum: 1
            maximum: 12
          description: Nombre maximal de lieux.
      responses:
        '200':
          description: Lieux trouvés.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlacesResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '502':
          $ref: '#/components/responses/IdfmUnavailable'
        '503':
          $ref: '#/components/responses/IdfmMissingToken'
        '504':
          $ref: '#/components/responses/IdfmTimeout'
  /api/idfm/journeys:
    get:
      tags:
        - IDF Mobilités
      summary: Calculer les itinéraires entre deux lieux.
      description: Renvoie des itinéraires marche, vélo et transports. Des trajets directs marche/vélo peuvent être complétés avec OSRM quand IDF Mobilités n'en fournit pas.
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
          description: Identifiant IDF Mobilités du départ.
        - name: to
          in: query
          required: true
          schema:
            type: string
          description: Identifiant IDF Mobilités de l'arrivée.
        - name: fromLon
          in: query
          required: false
          schema:
            type: number
          description: Longitude du départ pour les trajets directs.
        - name: fromLat
          in: query
          required: false
          schema:
            type: number
          description: Latitude du départ pour les trajets directs.
        - name: toLon
          in: query
          required: false
          schema:
            type: number
          description: Longitude de l'arrivée pour les trajets directs.
        - name: toLat
          in: query
          required: false
          schema:
            type: number
          description: Latitude de l'arrivée pour les trajets directs.
        - name: wheelchairAccessible
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: Limite les itinéraires aux transports accessibles en fauteuil roulant quand la valeur vaut true.
      responses:
        '200':
          description: Itinéraires calculés.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JourneysResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          description: Aucun itinéraire trouvé.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '502':
          $ref: '#/components/responses/IdfmUnavailable'
        '503':
          $ref: '#/components/responses/IdfmMissingToken'
        '504':
          $ref: '#/components/responses/IdfmTimeout'
  /api/idfm/bike-station-journey:
    post:
      tags:
        - IDF Mobilités
      summary: Composer un itinéraire à vélo via des bornes Vélib.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BikeStationJourneyRequest'
      responses:
        '200':
          description: Itinéraire composé.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BikeStationJourneyResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          description: Aucune borne de retour disponible.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '502':
          $ref: '#/components/responses/IdfmUnavailable'
        '504':
          $ref: '#/components/responses/IdfmTimeout'
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: urbanflow_auth
    csrfHeader:
      type: apiKey
      in: header
      name: x-csrf-token
  parameters:
    Lon:
      name: lon
      in: query
      required: true
      schema:
        type: number
        minimum: -180
        maximum: 180
      description: Longitude.
    Lat:
      name: lat
      in: query
      required: true
      schema:
        type: number
        minimum: -90
        maximum: 90
      description: Latitude.
  responses:
    ValidationError:
      description: Requête invalide.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentification requise ou session invalide.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    CsrfError:
      description: Jeton CSRF absent ou invalide.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequests:
      description: Trop de requêtes sur la fenêtre de limitation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServerError:
      description: Erreur serveur.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    IdfmMissingToken:
      description: Jeton IDF Mobilités manquant.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    IdfmUnavailable:
      description: API IDF Mobilités inaccessible.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    IdfmTimeout:
      description: Délai dépassé pour l'API IDF Mobilités.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        details:
          type: array
          items:
            type: string
    HealthResponse:
      type: object
      required:
        - status
        - service
      properties:
        status:
          type: string
          example: ok
        service:
          type: string
          example: UrbanFlow Mobility API
    AuthCredentials:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 12
          example: Password123!
    User:
      type: object
      required:
        - id
        - email
        - createdAt
      properties:
        id:
          oneOf:
            - type: string
              format: uuid
            - type: integer
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
    CsrfTokenResponse:
      type: object
      required:
        - csrfToken
      properties:
        csrfToken:
          type: string
          example: 8d51c4a4f4a52d73b0b88bda2d96834aa5f97b2d8d5f0c8f33ad94272adf9e2b.2f8d4f1d0fd0337eb31d2c7d9d8c7c4a6b729066391bb8f3c66a5bb1d31d8f21
    UserResponse:
      type: object
      required:
        - user
      properties:
        user:
          $ref: '#/components/schemas/User'
    CurrentUserResponse:
      type: object
      required:
        - user
      properties:
        user:
          nullable: true
          oneOf:
            - $ref: '#/components/schemas/User'
    Coordinates:
      type: array
      minItems: 2
      maxItems: 2
      items:
        type: number
      example:
        - 2.347
        - 48.858
    Line:
      type: object
      properties:
        id:
          type: string
          nullable: true
        code:
          type: string
          nullable: true
        label:
          type: string
          nullable: true
        commercialMode:
          type: string
          nullable: true
        physicalMode:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        textColor:
          type: string
          nullable: true
    Disruption:
      type: object
      properties:
        id:
          type: string
          nullable: true
        uri:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        type:
          type: string
          enum:
            - interruption
            - perturbation
        title:
          type: string
        message:
          type: string
        cause:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        severity:
          type: object
          properties:
            name:
              type: string
              nullable: true
            effect:
              type: string
              nullable: true
            color:
              type: string
              nullable: true
        applicationPeriods:
          type: array
          items:
            type: object
        line:
          $ref: '#/components/schemas/Line'
        severityRank:
          type: integer
        modeRank:
          type: integer
        count:
          type: integer
          description: Nombre de messages regroupes pour cette ligne.
        disruptions:
          type: array
          description: Detail des perturbations regroupees pour cette ligne.
          items:
            type: object
    Place:
      type: object
      required:
        - id
        - name
        - type
        - coordinates
        - lines
      properties:
        id:
          type: string
        label:
          type: string
        name:
          type: string
        type:
          type: string
        distance:
          type: number
          nullable: true
        coordinates:
          $ref: '#/components/schemas/Coordinates'
        city:
          type: string
          nullable: true
        lines:
          type: array
          items:
            $ref: '#/components/schemas/Line'
    BikeStation:
      allOf:
        - $ref: '#/components/schemas/Place'
        - type: object
          properties:
            stationId:
              type: string
            capacity:
              type: integer
            availableBikes:
              type: integer
            availableDocks:
              type: integer
            mechanicalBikes:
              type: integer
            electricBikes:
              type: integer
    BikeStationsResponse:
      type: object
      required:
        - stations
        - pagination
      properties:
        stations:
          type: array
          items:
            $ref: '#/components/schemas/BikeStation'
        pagination:
          type: object
          nullable: true
    NearbyStationsResponse:
      type: object
      required:
        - stations
        - pagination
      properties:
        stations:
          type: array
          items:
            $ref: '#/components/schemas/Place'
        pagination:
          type: object
          nullable: true
    PlacesResponse:
      type: object
      required:
        - places
        - pagination
      properties:
        places:
          type: array
          items:
            $ref: '#/components/schemas/Place'
        pagination:
          type: object
          nullable: true
    DisruptionsResponse:
      type: object
      required:
        - disruptions
        - pagination
      properties:
        disruptions:
          type: array
          items:
            $ref: '#/components/schemas/Disruption'
        pagination:
          type: object
          nullable: true
    JourneySection:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        mode:
          type: string
        label:
          type: string
        duration:
          type: integer
        from:
          type: string
          nullable: true
        to:
          type: string
          nullable: true
        departureDateTime:
          type: string
          nullable: true
        arrivalDateTime:
          type: string
          nullable: true
        color:
          type: string
        textColor:
          type: string
        line:
          allOf:
            - $ref: '#/components/schemas/Line'
          nullable: true
        distanceKm:
          type: number
        stopCount:
          type: integer
          nullable: true
        stops:
          type: array
          items:
            type: string
        geometry:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/Coordinates'
    CarbonFootprint:
      type: object
      properties:
        total_co2e:
          type: number
        unit:
          type: string
          example: g
        segments:
          type: array
          items:
            type: object
        car_solo_co2e:
          type: number
        savings_vs_car_solo_co2e:
          type: number
    CompletedJourney:
      type: object
      required:
        - id
        - type
        - completedAt
        - carbonFootprint
      properties:
        id:
          type: string
          maxLength: 120
        journeyId:
          type: string
          nullable: true
        type:
          type: string
          maxLength: 100
        completedAt:
          type: string
          format: date-time
        distanceKm:
          type: number
          nullable: true
        carbonFootprint:
          $ref: '#/components/schemas/CarbonFootprint'
    CompletedJourneyResponse:
      type: object
      required:
        - journey
      properties:
        journey:
          $ref: '#/components/schemas/CompletedJourney'
    CompletedJourneysResponse:
      type: object
      required:
        - journeys
      properties:
        journeys:
          type: array
          items:
            $ref: '#/components/schemas/CompletedJourney'
    Journey:
      type: object
      properties:
        id:
          type: string
        profile:
          type: string
          enum:
            - walking
            - bike
            - transit
        duration:
          type: integer
        walkingDuration:
          type: integer
        bikeDuration:
          type: integer
        nbTransfers:
          type: integer
        departureDateTime:
          type: string
          nullable: true
        arrivalDateTime:
          type: string
          nullable: true
        sections:
          type: array
          items:
            $ref: '#/components/schemas/JourneySection'
        geometry:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/Coordinates'
        carbonFootprint:
          $ref: '#/components/schemas/CarbonFootprint'
        bikeStations:
          type: object
          properties:
            start:
              $ref: '#/components/schemas/BikeStation'
            end:
              $ref: '#/components/schemas/BikeStation'
    JourneysResponse:
      type: object
      required:
        - journeys
        - carbonFootprintMessage
      properties:
        journeys:
          type: array
          items:
            $ref: '#/components/schemas/Journey'
        carbonFootprintMessage:
          type: string
          nullable: true
    BikeStationJourneyRequest:
      type: object
      required:
        - fromCoordinates
        - toCoordinates
        - startStation
      properties:
        fromCoordinates:
          $ref: '#/components/schemas/Coordinates'
        toCoordinates:
          $ref: '#/components/schemas/Coordinates'
        startStation:
          $ref: '#/components/schemas/BikeStation'
    BikeStationJourneyResponse:
      type: object
      required:
        - journey
        - carbonFootprintMessage
      properties:
        journey:
          $ref: '#/components/schemas/Journey'
        carbonFootprintMessage:
          type: string
          nullable: true